use for in instead of Object.keys for each in setAttributes and setUniforms
I prefer Object.keys for each but it's slower because it has to generate an
array. Maybe that will be fixed in ES5 or ES6 some how?

In any case for in usually requires checking hasOwnProperty but since
we're already checking the setters it's unlikely to ever be an issue