change so attribute prefix is "" by default.
I'm really mixed on this. I like `a_` as the prefix on my attributes
(and `u_` on uniforms and `v_` on varyings. I think it makes the GLSL
clearer because you know at a glance where the value is coming from.

But, when creating the geometry I prefer unprefixed as in

    position: [1, -1, 0, ...],
    normal: [0, 1, 0, ...],

etc. That means when going from geometry/buffers to attributes I need
a mapping.

I could decide to stop using the `a_` prefix in GLSL but I feel like
that's less readable. I could also have start prefixing geometry/buffers
but that feels gross too for some stupid reason

In the end though, rather than force my prefix on everyone I just made
it default to unprefixed. That will probably work with more people's
GLSL

There's still the case of unmatched names. For example your attribute
might be called `position` but your buffer is called `positions`.

For that case you have to specify a name

    var arrays = {
      positions: { attrib: "position", data: [1, -1, 0, ...], }
    };

Not much I can do about that case.