update build to deal with browserify
maybe there is a better way to do this. I'm not sure.

What happens is r+almond generates an AMD file. The problem
is if you try to browserify that browserify barfs because
it's doing some kind of text search for "require".

The solution seems to be munghing the file be replacing "require"
to something else like "foobarmoo". It's just an identifier
once wrapped with almond so it seems to work and browserify
is no longer unhappy.