hack node to test es6 modules
I really wish more was written around why node decided files must be named .mjs for node es6 modules.

They list the requirement that both `require` and `import` should be able to load both kinds of modules. Why that requirement? Why not `require` loads CJS and `import` loads ESM!??!

As it is, if you want to use modules in both node and the browser you're going to have to rename all your files to .mjs