Create a TypeScript test ("npm run testts")
"npm run test" will now run both the pre-existing "js" tests as well as the new "ts" test. Use "npm run testjs" to run only the earlier tests.

Borrows a handful of lines from code in the examples/ directory. It doesn't need to do anything interesting, it just needs to successfully compile.

The TS test is not very thorough at this point, but it's a very solid start imo.