ignore the .LICENSE files
AFAICT there the TerserPlugin now adds its own banner. To stop it
you have to use `extractComments` which seems to be the only way to
provide your own banner. BUT, extractComments always writes a file.
I could tell it `/dev/null` but that would fail on windows !?!?

If you know a solution send a PR. Otherwise I don't have time to waste
so just ignoring the files (took an 1 hour to fix this change already)