replace error with warning
getNumElementsFromAttributes, called from createBufferInfoFromArrays, throws an error when the buffer size does not evenly divide the attribute size. However, there are valid use cases for having an uneven division (e.g. reusing a large, preallocated webgl buffer to render different kinds of geometry). This commit changes the error to a warning, since this situation may be indicative of a bug in the user's code.