Corrected Sentence structure.
In the TL;DR section, the sentence is written as, " If you want to get stuff done use three.js. If you want to do stuff low-level with WebGL consider using TWGL."
Here in the first sentence, there must be a comma after the subordinate clause (done). 
The correct sentence could be written as, " If you want to get stuff done, use three.js."

The second sentence is structurally incorrect.
The correct sentence could be written as, " If you want to do low-level stuff with WebGL, consider using TWGL."
Thank you.